refactor(tests): refactor into split tests
* Each client is now tested separately.
* For clients that create resources, we now use a common pattern of
maintaining a multiset of resource identifiers to clean up at the end of
the test class.
* For operations that create resources, we now use the functionality to
wait on operations rather than sleeping.
* If an operation fails, we will fail the test (or setup)
