+++ src/test/java/com/google/cloud/compute/v1/it/ITInstanceGroupManagerTest.java
@@ -16,19 +16,18 @@
 package com.google.cloud.compute.v1.it;
 
 import static com.google.common.truth.Truth.assertThat;
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.fail;
 
 import com.google.api.gax.core.FixedCredentialsProvider;
 import com.google.auth.Credentials;
 import com.google.auth.oauth2.GoogleCredentials;
-import com.google.cloud.ServiceOptions;
 import com.google.cloud.compute.v1.AttachedDisk;
 import com.google.cloud.compute.v1.Disk;
 import com.google.cloud.compute.v1.DiskClient;
 import com.google.cloud.compute.v1.DiskSettings;
 import com.google.cloud.compute.v1.DiskTypeSettings;
-import com.google.cloud.compute.v1.GlobalOperationClient;
-import com.google.cloud.compute.v1.GlobalOperationSettings;
 import com.google.cloud.compute.v1.InstanceGroupManager;
 import com.google.cloud.compute.v1.InstanceGroupManagerClient;
 import com.google.cloud.compute.v1.InstanceGroupManagerSettings;
@@ -39,36 +38,27 @@ import com.google.cloud.compute.v1.InstanceTemplateClient;
 import com.google.cloud.compute.v1.InstanceTemplateSettings;
 import com.google.cloud.compute.v1.NetworkInterface;
 import com.google.cloud.compute.v1.Operation;
-import com.google.cloud.compute.v1.ProjectGlobalOperationName;
 import com.google.cloud.compute.v1.ProjectName;
 import com.google.cloud.compute.v1.ProjectZoneName;
-import com.google.cloud.compute.v1.ProjectZoneOperationName;
-import com.google.cloud.compute.v1.ZoneOperationClient;
-import com.google.cloud.compute.v1.ZoneOperationSettings;
 import com.google.common.collect.ArrayListMultimap;
 import com.google.common.collect.ListMultimap;
 import com.google.common.collect.Lists;
 import java.io.IOException;
 import java.util.List;
-import java.util.UUID;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-public class ITInstanceGroupManagerTest {
-  private static final String DEFAULT_PROJECT = ServiceOptions.getDefaultProjectId();
-  private static final ProjectName PROJECT_NAME = ProjectName.of(DEFAULT_PROJECT);
-  private static final String ZONE = "us-central1-a";
-  private static final String UID = UUID.randomUUID().toString();
-  private static final String ID = UID.substring(0, 8);
+public class ITInstanceGroupManagerTest extends BaseTest {
   private static final ProjectZoneName PROJECT_ZONE_NAME =
       ProjectZoneName.of(DEFAULT_PROJECT, ZONE);
+  private static final String DISK_NAME =
+      TestHelper.getTestUniqueName("instance-group-manager-disk");
+  private static final String INSTANCE_GROUP_NAME = TestHelper.getTestUniqueName("instance-group");
 
   private static DiskClient diskClient;
   private static InstanceGroupManagerClient instanceGroupManagerClient;
   private static InstanceTemplateClient instanceTemplateClient;
-  private static GlobalOperationClient globalOperationClient;
-  private static ZoneOperationClient zoneOperationClient;
 
   private static ListMultimap<String, String> resourcesToCleanUp = ArrayListMultimap.create();
 
@@ -95,30 +85,22 @@ public class ITInstanceGroupManagerTest {
             .build();
     instanceGroupManagerClient = InstanceGroupManagerClient.create(instanceGroupManagerSettings);
 
-    GlobalOperationSettings globalOperationSettings =
-        GlobalOperationSettings.newBuilder().setCredentialsProvider(credentialsProvider).build();
-    globalOperationClient = GlobalOperationClient.create(globalOperationSettings);
-
-    ZoneOperationSettings zoneOperationSettings =
-        ZoneOperationSettings.newBuilder().setCredentialsProvider(credentialsProvider).build();
-    zoneOperationClient = ZoneOperationClient.create(zoneOperationSettings);
-
     // create a disk
-    Disk disk = Disk.newBuilder().setName("test-" + ID).build();
+    Disk disk = Disk.newBuilder().setName(DISK_NAME).build();
     Operation completedOperation = waitForOperation(diskClient.insertDisk(PROJECT_ZONE_NAME, disk));
     resourcesToCleanUp.put("disk", completedOperation.getTargetLink());
 
     // create an instance template
     InstanceTemplate instanceTemplate =
         InstanceTemplate.newBuilder()
-            .setName("test-" + ID)
+            .setName(INSTANCE_GROUP_NAME)
             .setProperties(
                 InstanceProperties.newBuilder()
                     .setMachineType("n1-standard-1")
                     .addDisks(
                         AttachedDisk.newBuilder()
                             .setType("PERSISTENT")
-                            .setSource("test-" + ID)
+                            .setSource(DISK_NAME)
                             .setBoot(true)
                             .build())
                     .addNetworkInterfaces(
@@ -135,8 +117,8 @@ public class ITInstanceGroupManagerTest {
     // Create an instance group manager
     InstanceGroupManager instanceGroupManager =
         InstanceGroupManager.newBuilder()
-            .setBaseInstanceName("base-instance-" + ID)
-            .setName("test-instance-group-" + ID)
+            .setBaseInstanceName(INSTANCE_GROUP_NAME)
+            .setName(INSTANCE_GROUP_NAME)
             .setTargetSize(0)
             .setInstanceTemplate(instanceTemplateSelfLink)
             .build();
@@ -164,8 +146,6 @@ public class ITInstanceGroupManagerTest {
     diskClient.close();
     instanceGroupManagerClient.close();
     instanceTemplateClient.close();
-    globalOperationClient.close();
-    zoneOperationClient.close();
   }
 
   @Test
@@ -216,14 +196,5 @@ public class ITInstanceGroupManagerTest {
       }
     }
     fail("expected to find the inserted instance group manager");
-  }
-
-  private static Operation waitForOperation(Operation operation) {
-    if (operation.getZone() != null) {
-      return zoneOperationClient.waitZoneOperation(
-          ProjectZoneOperationName.of(operation.getName(), DEFAULT_PROJECT, ZONE));
-    }
-    return globalOperationClient.waitGlobalOperation(
-        ProjectGlobalOperationName.of(operation.getName(), DEFAULT_PROJECT));
   }
 }
