+++ src/test/java/com/google/cloud/compute/v1/it/ITInstanceTemplateTest.java
@@ -1,0 +1,53 @@
+/*
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.google.cloud.compute.v1.it;
+
+import static com.google.common.truth.Truth.assertThat;
+
+import com.google.cloud.compute.v1.InstanceTemplate;
+import com.google.cloud.compute.v1.InstanceTemplateClient;
+import com.google.cloud.compute.v1.InstanceTemplateSettings;
+import com.google.common.collect.Lists;
+import java.io.IOException;
+import java.util.List;
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
+import org.junit.Test;
+
+public class ITInstanceTemplateTest extends BaseTest {
+
+  private static InstanceTemplateClient instanceTemplateClient;
+
+  @BeforeClass
+  public static void setUp() throws IOException {
+    InstanceTemplateSettings instanceTemplateSettings =
+        InstanceTemplateSettings.newBuilder().setCredentialsProvider(credentialsProvider).build();
+    instanceTemplateClient = InstanceTemplateClient.create(instanceTemplateSettings);
+  }
+
+  @AfterClass
+  public static void tearDown() {
+    instanceTemplateClient.close();
+  }
+
+  @Test
+  public void listInstanceTemplatesTest() {
+    List<InstanceTemplate> templates =
+        Lists.newArrayList(instanceTemplateClient.listInstanceTemplates(PROJECT_NAME).iterateAll());
+    assertThat(templates).isNotNull();
+    assertThat(templates.contains(null)).isFalse();
+  }
+}
