+++ src/test/java/com/google/cloud/compute/v1/it/ITNodeGroupTest.java
@@ -1,0 +1,65 @@
+/*
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.google.cloud.compute.v1.it;
+
+import static com.google.common.truth.Truth.assertThat;
+
+import com.google.cloud.compute.v1.NodeGroup;
+import com.google.cloud.compute.v1.NodeGroupClient;
+import com.google.cloud.compute.v1.NodeGroupSettings;
+import com.google.cloud.compute.v1.NodeGroupsScopedList;
+import com.google.common.collect.Lists;
+import java.io.IOException;
+import java.util.List;
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
+import org.junit.Test;
+
+public class ITNodeGroupTest extends BaseTest {
+
+  private static NodeGroupClient nodeGroupClient;
+
+  @BeforeClass
+  public static void setUp() throws IOException {
+    NodeGroupSettings nodeGroupSettings =
+        NodeGroupSettings.newBuilder().setCredentialsProvider(credentialsProvider).build();
+    nodeGroupClient = NodeGroupClient.create(nodeGroupSettings);
+  }
+
+  @AfterClass
+  public static void tearDown() {
+    nodeGroupClient.close();
+  }
+
+  @Test
+  public void listNodeGroupsTest() {
+    List<NodeGroup> nodeGroups =
+        Lists.newArrayList(nodeGroupClient.listNodeGroups(PROJECT_ZONE_NAME).iterateAll());
+    assertThat(nodeGroups).isNotNull();
+    assertThat(nodeGroups.size()).isEqualTo(0);
+    assertThat(nodeGroups.contains(null)).isFalse();
+  }
+
+  @Test
+  public void aggregatedListNodeGroupsTest() {
+    List<NodeGroupsScopedList> nodeGroupsScopedLists =
+        Lists.newArrayList(
+            nodeGroupClient.aggregatedListNodeGroups(true, PROJECT_NAME).iterateAll());
+    assertThat(nodeGroupsScopedLists).isNotNull();
+    assertThat(nodeGroupsScopedLists.size()).isGreaterThan(0);
+    assertThat(nodeGroupsScopedLists.contains(null)).isFalse();
+  }
+}
