+++ src/test/java/com/google/cloud/compute/v1/it/ITVpnGatewayTest.java
@@ -1,0 +1,51 @@
+/*
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.google.cloud.compute.v1.it;
+
+import static com.google.common.truth.Truth.assertThat;
+
+import com.google.cloud.compute.v1.VpnGateway;
+import com.google.cloud.compute.v1.VpnGatewayClient;
+import com.google.cloud.compute.v1.VpnGatewaySettings;
+import java.io.IOException;
+import java.util.Iterator;
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
+import org.junit.Test;
+
+public class ITVpnGatewayTest extends BaseTest {
+
+  private static VpnGatewayClient vpnGatewayClient;
+
+  @BeforeClass
+  public static void setUp() throws IOException {
+    VpnGatewaySettings vpnGatewaySettings =
+        VpnGatewaySettings.newBuilder().setCredentialsProvider(credentialsProvider).build();
+    vpnGatewayClient = VpnGatewayClient.create(vpnGatewaySettings);
+  }
+
+  @AfterClass
+  public static void tearDown() {
+    vpnGatewayClient.close();
+  }
+
+  @Test
+  public void listVpnGatewaysTest() {
+    Iterator<VpnGateway> vpnGatewayIterator =
+        vpnGatewayClient.listVpnGateways(PROJECT_REGION_NAME).iterateAll().iterator();
+    assertThat(vpnGatewayIterator).isNotNull();
+  }
+}
