+++ src/test/java/com/google/cloud/compute/v1/it/TestHelper.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.google.cloud.compute.v1.it;
+
+import java.util.UUID;
+
+public class TestHelper {
+  private static final String UID = UUID.randomUUID().toString();
+  private static final String ID = UID.substring(0, 8);
+
+  public static String getTestUniqueName(String basename) {
+    return "test-" + basename + "-" + ID;
+  }
+
+  public static String getUid() {
+    return UID;
+  }
+}
